% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{clear_meetupr_token}
\alias{clear_meetupr_token}
\title{Remove all credentials and cached tokens for a client}
\usage{
clear_meetupr_token(client_name = get_client_name())
}
\arguments{
\item{client_name}{OAuth client name}
}
\description{
Remove all credentials and cached tokens for a client
}
