% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup-query.R
\docType{data}
\name{events_query}
\alias{events_query}
\title{Clean Query Definitions Using Shared Infrastructure
This file completely replaces the existing R/meetup-query.R}
\format{
An object of class \code{meetupr::MeetupQuery} (inherits from \code{S7_object}) of length 1.
}
\usage{
events_query
}
\description{
Clean Query Definitions Using Shared Infrastructure
This file completely replaces the existing R/meetup-query.R
}
\keyword{internal}
