% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql-introspection.R
\name{explore_mutations}
\alias{explore_mutations}
\title{Explore available mutations in the Meetup GraphQL API}
\usage{
explore_mutations(schema = meetup_introspect())
}
\arguments{
\item{schema}{The schema object obtained from \code{meetup_introspect()}. If NULL,
the function will call \code{meetup_introspect()} to get the schema.}
}
\value{
A tibble with details about each mutation, including mutation name,
description, argument count, and return type. If no mutations are
available,
}
\description{
This function retrieves and displays the available mutations in the Meetup
GraphQL API schema.
}
\examples{
\dontrun{
explore_mutations()
}
}
