% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql-introspection.R
\name{explore_query_fields}
\alias{explore_query_fields}
\title{Explore available query fields in the Meetup GraphQL API}
\usage{
explore_query_fields(schema = meetup_introspect())
}
\arguments{
\item{schema}{The schema object obtained from \code{meetup_introspect()}. If NULL,
the function will call \code{meetup_introspect()} to get the schema.}
}
\value{
A tibble with details about each query field, including field name,
description, argument count, and return type.
}
\description{
This function retrieves and displays the available query fields in the Meetup
GraphQL API schema.
}
\examples{
\dontrun{
explore_query_fields()
}
}
