% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_result_dataframe.R
\name{extract_result_dataframe}
\alias{extract_result_dataframe}
\title{Extracts and flattens/unnests a data.frame from a JSON response}
\usage{
extract_result_dataframe(l)
}
\arguments{
\item{l}{A list of JSON responses of the same kind}
}
\value{
A dataframe
}
\description{
Extracts and flattens/unnests a data.frame from a JSON response
}

