% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_events.R
\name{find_events}
\alias{find_events}
\title{Get locations for meetup groups}
\usage{
find_events(
  query = NULL,
  topic_category = NULL,
  lat = NULL,
  lon = NULL,
  radius = NULL,
  start_date_range = NULL,
  end_date_range = NULL,
  fields = NULL,
  excluded_groups = NULL,
  order = NULL,
  api_key = NULL
)
}
\arguments{
\item{lat}{Latitude (optional)}

\item{lon}{Longitude (optional)}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{

}
\description{
Provide either a search query or a latitude + longitude combination to find a location
}
\examples{
\dontrun{
ll <- get_locations("New York City")
find_events(ll$lat, ll$lon, radius = 10, query = "ultimate frisbee")

topics <- get_topics()
mvmt_ids <- topics \%>\% dplyr::filter(name == "Movements") \%>\% dplyr::pull(category_ids)
find_events(ll$lat, ll$lon, radius = 10, topic_category = mvmt_ids)
}
}
