% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_upcoming.R
\name{find_upcoming}
\alias{find_upcoming}
\title{Meetup search for upcoming events}
\usage{
find_upcoming(search, start_date, long, lati, .radius = 10, .page = 50,
  ...)
}
\arguments{
\item{search}{Text to use to search events as a single character value}

\item{start_date}{Start date in the form}

\item{long}{Longitude of point to search near}

\item{lati}{Latitude of point to search near}

\item{.radius}{Number of miles of search radius around lat, long point to find events. Default is 10.}

\item{.page}{Maximum number of results to return. Default is 50.}

\item{...}{Additional parameters for API call. See references.}
}
\description{
Meetup search for upcoming events
}
\examples{

}
\references{
\url{https://secure.meetup.com/meetup_api/console/?path=/find/upcoming_events}
}
