% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url_request.R
\name{generate_url_request}
\alias{generate_url_request}
\title{Generates the URL for the API call}
\usage{
generate_url_request(url_prefix = "https://api.meetup.com/2/", service,
  key = NULL, group_urlname = NULL, group_id = NULL, event_id = NULL,
  page = NULL, offset = NULL, order = NULL, authentication = "oauth")
}
\value{
URL for API call
}
\description{
Generates the URL for the API call
}

