% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_process_group_data.R
\name{get_and_process_group_data}
\alias{get_and_process_group_data}
\title{Retrieves json objects for group membership, event and rsvp information
and creates dataframes}
\usage{
get_and_process_group_data(group_urlname, meetup_token)
}
\arguments{
\item{group_urlname}{The group name as shown in the meetup.com URL}

\item{meetup_token}{The token object}
}
\value{
list of data.frames for group, members, profiles, events, and time
}
\description{
Retrieves json objects for group membership, event and rsvp information
and creates dataframes
}

