% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_auth.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the legacy API key from the .state environment}
\usage{
get_api_key()
}
\description{
Get the legacy API key from the .state environment
}
\keyword{internal}
