% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boards.R
\name{get_boards}
\alias{get_boards}
\title{Get the discussion boards from a meetup group}
\usage{
get_boards(urlname, verbose = meetupr_verbose())
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item description
\item created
\item updated
\item post_count
\item discussion_count
\item latest_reply_created
\item latest_reply_member_name
\item resource
}
}
\description{
Get the discussion boards from a meetup group
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
meetup_boards <- get_boards(urlname = urlname)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/:urlname/boards/}
}
