% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_client_name}
\alias{get_client_name}
\title{Get Meetup OAuth Client Name}
\usage{
get_client_name()
}
\value{
A string representing the client name.
}
\description{
Retrieves the Meetup OAuth client name from the
\code{MEETUPR_CLIENT_NAME} environment variable or defaults to "meetupr".
}
