% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_group.R
\name{get_data_for_group}
\alias{get_data_for_group}
\title{Retrieves json objects for group membership, event and rsvp information}
\usage{
get_data_for_group(group_urlname, meetup_token)
}
\arguments{
\item{group_urlname}{The group name as shown in the meetup.com URL}

\item{meetup_token}{The token object}
}
\value{
list of responses for group, members, profiles, events, and time
}
\description{
Retrieves json objects for group membership, event and rsvp information
}

