% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-events.R
\name{get_events}
\alias{get_events}
\title{Get the events from a meetup group}
\usage{
get_events(urlname, status = NULL, ..., extra_graphql = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{status}{Character vector of event statuses to retrieve.}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\description{
Get the events from a meetup group
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_events", package = "meetupr")
meetupr:::mock_if_no_auth()
}
get_events("rladies-lagos", "past")
\dontshow{
vcr::eject_cassette()
}
}
