% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get list of meetup groups using text-based search}
\usage{
get_groups(text = NULL, radius = "global", api_key = NULL)
}
\arguments{
\item{text}{string used to filter groups}

\item{radius}{can be either "global" (default) or distance in miles in the
range 0-100.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
Tibble containing the list of selected groups. The table contains 12
columns: name, urlname, who, members, status, created, organizer, lat, lon,
city, country, timezone.
}
\description{
Get list of meetup groups using text-based search
}
\examples{
\dontrun{
api_key <- Sys.getenv("rladies_api_key")
groups <- get_groups(text = "r-ladies", api_key = api_key)
}
}
