% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get locations for meetup groups}
\usage{
get_locations(query = NULL, lat = NULL, lon = NULL, api_key = NULL)
}
\arguments{
\item{query}{Search query: a city/town name or zip code}

\item{lat}{Latitude (optional)}

\item{lon}{Longitude (optional)}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{

}
\description{
Provide either a search query or a latitude + longitude combination to find a location
}
\examples{
\dontrun{
get_locations("New York City")
get_locations("12345")
}
}
