% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetup_attendees.R
\name{get_meetup_attendees}
\alias{get_meetup_attendees}
\title{Get the attendees}
\usage{
get_meetup_attendees(urlname, api_key, event_id)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}

\item{event_id}{The id of the event.}
}
\description{
Get the attendees
}

