% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetup_token.R
\name{get_meetup_token}
\alias{get_meetup_token}
\title{Obtains the oath1.0 token required to get data from meetup.com}
\usage{
get_meetup_token(key, secret)
}
\arguments{
\item{key}{Your meetup.com oauth key}

\item{secret}{Your meetup.com oauth secret key}
}
\value{
token object required for other functions
}
\description{
Obtains the oath1.0 token required to get data from meetup.com
}

