% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Get the current meetup members}
\usage{
get_members(urlname, api_key)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}
}
\description{
Get the current meetup members
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
api_key <- Sys.getenv("rladies_api_key")
members <- get_members(urlname, api_key)
}
}

