% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{get_pro_events}
\alias{get_pro_events}
\title{Get the events from a PRO meetup group}
\usage{
get_pro_events(urlname, verbose = TRUE)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item created
\item status
\item time
\item local_date
\item local_time
\item waitlist_count
\item yes_rsvp_count
\item venue_id
\item venue_name
\item venue_lat
\item venue_lon
\item venue_address_1
\item venue_city
\item venue_state
\item venue_zip
\item venue_country
\item description
\item link
\item resource
}
}
\description{
This can only fetch events for the
next 30 days.
}
\examples{
\dontrun{
urlname <- "rladies"
past_events <- get_events(urlname = urlname,
                      event_status = "past")
upcoming_events <- get_events(urlname = urlname,
                      event_status = "upcoming")
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/:urlname/events/#list}
}
