% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{get_pro_groups}
\alias{get_pro_groups}
\title{Get the current meetup members from a pro meetup group}
\usage{
get_pro_groups(
  urlname,
  verbose = getOption("meetupr.verbose", rlang::is_interactive())
)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item status
\item founded
\item members
\item upcoming_events
\item past events
\item city
\item country
\item state
\item lat
\item lon
\item urlname
}
}
\description{
Get the current meetup members from a pro meetup group
}
\examples{
\dontrun{
urlname <- "rladies"
members <- get_pro_groups(urlname)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/pro/:urlname/groups/}
}
