% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real_time.R
\name{get_real_time}
\alias{get_real_time}
\title{Converts the meetup time value to a POSIXct value in EST}
\usage{
get_real_time(meetup_time)
}
\arguments{
\item{meetup_time}{The the timestamp value provided by meetup}
}
\value{
POSIXct object with EST timezone
}
\description{
Converts the meetup time value to a POSIXct value in EST
}

