% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{get_response}
\alias{get_response}
\title{Makes the GET call to the meetup API}
\usage{
get_response(service, group_urlname = NULL, meetup_token, default_ct = 800,
  group_id = NULL, order = NULL, event_id = NULL, offset = NULL)
}
\value{
Response object for the URL
}
\description{
Makes the GET call to the meetup API
}

