% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topics.R
\name{get_topics}
\alias{get_topics}
\title{Get topic categories for meetup groups}
\usage{
get_topics(lat = NULL, lon = NULL, fields = NULL, api_key = NULL)
}
\arguments{
\item{lat}{Latitude (optional)}

\item{lon}{Longitude (optional)}

\item{fields}{Character, character vector or characters separated by comma of optional fields to append to the response}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{

}
\description{
Get topic categories for meetup groups
}
\examples{
\dontrun{
get_topics()
}
}
