% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{has_jwt_credentials}
\alias{has_jwt_credentials}
\title{Check for JWT Credentials}
\usage{
has_jwt_credentials()
}
\value{
A logical value:
\itemize{
\item \code{TRUE} if all the required JWT credentials are set,
\item \code{FALSE} otherwise.
}
}
\description{
This function checks if the necessary environment variables for using
JWT authentication with Meetup API are set.
}
\details{
The function validates the existence of the following environment variables:
\itemize{
\item MEETUP_CLIENT_ID: Your Meetup client ID.
\item MEETUP_MEMBER_ID: Your Meetup member ID.
\item MEETUP_RSA_KEY or MEETUP_RSA_PATH: RSA key or its path used for signing.
}
}
\examples{
Sys.setenv(MEETUP_CLIENT_ID = "12345",
           MEETUP_MEMBER_ID = "67890",
           MEETUP_RSA_KEY = "example_rsa_key")
has_jwt_credentials() # Should return TRUE

Sys.setenv(MEETUP_RSA_KEY = "")
has_jwt_credentials() # Should return FALSE
}
