% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{has_oauth_credentials}
\alias{has_oauth_credentials}
\title{Check for OAuth Credentials}
\usage{
has_oauth_credentials()
}
\value{
A logical value:
\itemize{
\item \code{TRUE} if all the required OAuth credentials are set,
\item \code{FALSE} otherwise.
}
}
\description{
This function checks if the necessary environment variables for using
OAuth authentication with Meetup API are set.
}
\details{
The function validates the existence of the following environment variables:
\itemize{
\item MEETUP_CLIENT_ID: Your Meetup client ID.
\item MEETUP_CLIENT_SECRET: Your Meetup client secret.
}
}
\examples{
Sys.setenv(MEETUP_CLIENT_ID = "12345",
           MEETUP_CLIENT_SECRET = "example_secret")
has_oauth_credentials() # Should return TRUE

Sys.setenv(MEETUP_CLIENT_SECRET = "")
has_oauth_credentials() # Should return FALSE
}
