% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_meetupr_debug}
\alias{local_meetupr_debug}
\title{Temporarily enable debug mode}
\usage{
local_meetupr_debug(verbosity = 0, env = rlang::caller_env())
}
\arguments{
\item{verbosity}{How much debug information to show.
0 = off, 1 = basic, 2 = verbose, 3 = very verbose}

\item{env}{The environment in which to set the variable.}
}
\value{
The old debug value (invisibly)
}
\description{
Sets the level of verbosity for \code{httr2::local_verbosity()},
and the MEETUPR_DEBUG environment variable.
Can help debug issues with API requests.
}
\examples{
\dontrun{
local_meetupr_debug(2)

# Turn off debug mode
local_meetupr_debug(0)
}
}
