% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_auth.R
\name{meetup_auth}
\alias{meetup_auth}
\alias{meetup_token}
\alias{meetup_deauth}
\alias{meetup_token_path}
\title{Authorize \code{meetupr}}
\usage{
meetup_auth(
  token = meetup_token_path(),
  new_user = FALSE,
  key = getOption("meetupr.consumer_key"),
  secret = getOption("meetupr.consumer_secret"),
  cache = getOption("meetupr.httr_oauth_cache", TRUE),
  verbose = getOption("meetupr.verbose", rlang::is_interactive()),
  use_appdir = TRUE,
  token_path = NULL
)

meetup_token(verbose = FALSE)

meetup_deauth(
  clear_cache = TRUE,
  verbose = getOption("meetupr.verbose", rlang::is_interactive())
)

meetup_token_path()
}
\arguments{
\item{token}{optional; an actual token object or the path to a valid token
stored as an \code{.rds} file.}

\item{new_user}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
want to wipe the slate clean and re-authenticate with the same or different
Meetup account. This disables the current token that's at \code{meetup_token_path()}.}

\item{key, secret}{the "Client ID" and "Client secret" for the application;
defaults to the ID and secret built into the \code{meetupr} package}

\item{cache}{logical indicating if \code{meetupr} should cache
credentials a file.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}

\item{use_appdir}{Logical indicating whether to save the created token
in app dir as determined by \code{rappdirs::user_data_dir("meetupr", "meetupr")}.
If \code{cache} is \code{FALSE} this is ignored.}

\item{token_path}{Path where to save the token. If \code{use_appdir} is \code{TRUE},
this is ignored.}

\item{clear_cache}{logical indicating whether to disable the
token file, if such exists, by apprending "-SUSPENDED" to its name}
}
\value{
a \code{request} object (an S3 class provided by \code{httr})

Either NULL or the path in which the token is saved.
}
\description{
Authorize \code{meetupr} via the OAuth API. You will be directed to a web
browser, asked to sign in to your Meetup account, and to grant \code{meetupr}
permission to operate on your behalf. By default, these user credentials are
saved to an app dir as determined by \code{rappdirs::user_data_dir("meetupr", "meetupr")}.
If you set \code{use_appdir} to \code{FALSE} but \code{cache} to \code{TRUE},
they are cached in a file named \code{.httr-oauth} in the current working directory.
To control where the file is saved, use \code{token_path}.

If token is not already available, call \code{\link{meetup_auth}} to either
load from cache or initiate OAuth2.0 flow. Return the token -- not "bare"
but, rather, prepared for inclusion in downstream requests.

Suspend \code{\link{meetupr}}'s authorization to place requests to the Meetup
APIs on behalf of the authenticated user.
}
\section{How to force meetupr to use a given token?}{


Save this token somewhere on disk (\code{token_path} argument of \code{meetup_auth}).
Then in future sessions call \code{meetup_auth()} with \code{token} set to that path.
}

\section{Advanced usage}{


Most users, most of the time, do not need to call this function explicitly --
it will be triggered by the first action that requires authorization. Even
when called, the default arguments will often suffice. However, when
necessary, this function allows the user to

\itemize{
\item force the creation of a new token
\item retrieve current token as an object, for possible storage to an
\code{.rds} file
\item read the token from an object or from an \code{.rds} file
\item provide your own app key and secret -- this requires setting up
a new OAuth consumer on \href{https://secure.meetup.com/meetup_api/oauth_consumers/}{Meetup}
\item prevent caching of credentials
}

In a direct call to \code{meetup_auth}, the user can provide the token, app
key and secret explicitly and can dictate whether interactively-obtained
credentials will be cached. If unspecified, these
arguments are controlled via options, which, if undefined at the time
\code{meetupr} is loaded, are defined like so:

\describe{
\item{key}{Set to option \code{meetupr.consumer_key}, which defaults to a
consumer key that ships with the package}
\item{secret}{Set to option \code{meetupr.consumer_secret}, which defaults to
a consumer secret that ships with the package}
}

To override these defaults in persistent way, predefine one or more of them
with lines like this in a \code{.Rprofile} file:
\preformatted{
options(meetupr.consumer_key = "FOO",
        meetupr.consumer_secret = "BAR")
}
See \code{\link[base]{Startup}} for possible locations for this file and the
implications thereof.

More detail is available from
\href{https://www.meetup.com/meetup_api/auth/#oauth2-resources}{Authenticating
with the Meetup API}.
}

\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
meetup_auth()

## store token in an object and then to file
ttt <- meetup_auth()
saveRDS(ttt, "ttt.rds")

## load a pre-existing token
meetup_auth(token = ttt)       # from an object
meetup_auth(token = "ttt.rds") # from .rds file
}
\dontrun{
meetup_token()
}
\dontrun{
meetup_deauth()
}
meetup_token_path()
}
\concept{auth functions}
\keyword{internal}
