% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetup_auth}
\alias{meetup_auth}
\alias{meetup_deauth}
\title{Meetup API Authentication}
\usage{
meetup_auth(...)

meetup_deauth(
  client_name = Sys.getenv("MEETUP_CLIENT_NAME", "meetupr"),
  service = "meetupr",
  clear_keyring = TRUE
)
}
\arguments{
\item{...}{Additional arguments to \code{meetup_client()}.}

\item{client_name}{A string representing the name of the client. By
default, it is set to \code{"meetupr"} and retrieved from the
\code{MEETUP_CLIENT_NAME} environment variable.}

\item{service}{The name of the keyring service to clear.
Defaults to \code{"meetupr"}.}

\item{clear_keyring}{A logical value indicating whether to clear
the associated keyring entries. Defaults to \code{TRUE}.}
}
\value{
Nothing. Outputs messages indicating the result of the
process.
}
\description{
Functions to manage authentication with the Meetup API.
Includes functions to authenticate,
and deauthorize by removing cached credentials.
}
\section{Functions}{
\itemize{
\item \code{meetup_auth()}: Authenticate and display the
authenticated user's name.

\item \code{meetup_deauth()}: Remove cached authentication
for the Meetup API client.

}}
\examples{
\dontrun{
# Default deauthorization
meetup_deauth()

# Deauthorization with a custom client name
meetup_deauth(client_name = "custom_client")
}
}
