% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetup_auth_load_ci}
\alias{meetup_auth_load_ci}
\title{Load Meetup Authentication Token from CI Environment}
\usage{
meetup_auth_load_ci(client_name = Sys.getenv("MEETUP_CLIENT_NAME", "meetupr"))
}
\arguments{
\item{client_name}{A string representing the name of the client. By
default, it is set to \code{"meetupr"} and retrieved from the
\code{MEETUP_CLIENT_NAME} environment variable.}
}
\value{
Invisible \code{TRUE} if the token was successfully loaded and decoded.
If an error occurs, the function will throw an error and halt execution.
}
\description{
This function reads and decodes a Meetup API authentication token from
environment variables. The decoded token is then saved to a specified token
file within the OAuth cache directory for use by the application. This
function is designed to integrate with Continuous Integration (CI)
pipelines.
}
\details{
The function requires the following environment variables to be
set in the CI environment:
\itemize{
\item \code{MEETUP_TOKEN}: The base64-encoded token string.
\item \code{MEETUP_TOKEN_FILE}: The name of the file where the decoded token
will be stored.
}

If any of these variables are not set, the function will terminate with an
error.

The decoded token is saved in the OAuth cache directory at a path
constructed as \code{{oauth_cache_path}/{client_name}/{MEETUP_TOKEN_FILE}}.
}
\examples{
\dontrun{
  # Example usage in a CI pipeline with the required environment variables:
  meetup_auth_load_ci()

  # Overriding the default client name:
  meetup_auth_load_ci(client_name = "my_custom_client")
}

}
