% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetup_client}
\alias{meetup_client}
\title{Create a Meetup OAuth Client}
\usage{
meetup_client(
  client_id = Sys.getenv("MEETUP_CLIENT_ID"),
  client_secret = Sys.getenv("MEETUP_CLIENT_SECRET"),
  client_name = Sys.getenv("MEETUP_CLIENT_NAME", "meetupr"),
  ...
)
}
\arguments{
\item{client_id}{A string representing the Meetup client ID. By default,
it is retrieved from the \code{MEETUP_CLIENT_ID} environment variable.}

\item{client_secret}{A string representing the Meetup client secret. By
default, it is retrieved from the \code{MEETUP_CLIENT_SECRET} environment
variable.}

\item{client_name}{A string representing the name of the client. By
default, it is set to \code{"meetupr"} and retrieved from the
\code{MEETUP_CLIENT_NAME} environment variable.}

\item{...}{Additional arguments passed to the \code{httr2::oauth_client} function.}
}
\value{
An OAuth client object created with the \code{httr2::oauth_client}
function. This client can be used to handle authentication with the
Meetup API.
}
\description{
This function initializes and returns an OAuth client for authenticating
with the Meetup API. It requires the Meetup client ID and secret, which
can be passed as arguments or retrieved from environment variables.
}
\details{
If the \code{client_id} or \code{client_secret} parameters are empty, the function
will throw an error prompting you to set the \code{MEETUP_CLIENT_ID} and
\code{MEETUP_CLIENT_SECRET} environment variables.
}
\examples{
\dontrun{
# Example 1: Using environment variables to set credentials
client <- meetup_client()

# Example 2: Passing client ID and secret as arguments
client <- meetup_client(
  client_id = "your_client_id",
  client_secret = "your_client_secret"
)
}

}
