% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetup_deauth}
\alias{meetup_deauth}
\title{Deauthorize Meetup API Client}
\usage{
meetup_deauth(client_name = Sys.getenv("MEETUP_CLIENT_NAME", "meetupr"))
}
\arguments{
\item{client_name}{A string representing the name of the client. By
default, it is set to \code{"meetupr"} and retrieved from the
\code{MEETUP_CLIENT_NAME} environment variable.}
}
\value{
Returns an invisible logical value:
\itemize{
\item \code{TRUE} if the authentication cache was successfully removed.
\item \code{FALSE} if no authentication cache was found to remove.
}
}
\description{
This function removes the cached authentication for the
Meetup API client, effectively deauthorizing the client.
}
\details{
This function checks for the existence of a directory containing
cached authentication credentials. If found, the directory is deleted,
removing the cached authentication for the specified client. If the
directory does not exist, a message is displayed and the function returns
\code{FALSE}.
}
\examples{
\dontrun{
# Default deauthorization
meetup_deauth()

# Deauthorization with a custom client name
meetup_deauth(client_name = "custom_client")
}
}
