% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_pro.R
\name{get_pro_groups}
\alias{get_pro_groups}
\alias{get_pro_events}
\title{Meetup pro functions}
\usage{
get_pro_groups(
  urlname,
  verbose = getOption("meetupr.verbose", rlang::is_interactive())
)

get_pro_events(
  urlname,
  verbose = getOption("meetupr.verbose", rlang::is_interactive())
)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}
}
\value{
A tibble with meetup information
}
\description{
The pro functions only work if the querying users
had a meetup pro account.
}
\details{
\describe{
\item{get_pro_groups}{Get the current meetup members from a pro meetup group}
\item{get_pro_events}{Get pro group events for the enxt 30 days}
}
}
\examples{
\dontrun{
urlname <- "rladies"
members <- get_pro_groups(urlname)

past_events <- get_events(urlname = urlname,
                      event_status = "past")
upcoming_events <- get_events(urlname = urlname,
                      event_status = "upcoming")
}

}
\references{
\url{https://www.meetup.com/meetup_api/docs/pro/:urlname/groups/}
\url{https://www.meetup.com/meetup_api/docs/:urlname/events/#list}
}
