% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_query.R
\name{meetup_query}
\alias{meetup_query}
\title{General Meetup query}
\usage{
meetup_query(query, ..., token = meetup_token())
}
\arguments{
\item{query}{Required query text. Test queries at: https://www.meetup.com/api/playground}

\item{...}{Should be empty. Used for parameter expansion}

\item{token}{Meetup token}
}
\value{
raw list results from the query, not tidied
}
\description{
Create your own query to the Meetup API
}
\examples{
\dontrun{
 res <- meetup_query('
   query($eventId: ID = "103349942!chp") {
    event(id: $eventId) {
     title
     description
     dateTime
    }
   }
')
}

}
