% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetupr-package.R
\docType{package}
\name{meetupr-package}
\alias{meetupr}
\alias{meetupr-package}
\title{meetupr: Access Meetup Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides programmatic access to the 'Meetup' 'GraphQL' API (\url{https://www.meetup.com/api/schema/}), enabling users to retrieve information about groups, events, and members from 'Meetup' (\url{https://www.meetup.com/}). Supports authentication via 'OAuth2' and includes functions for common queries and data manipulation tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rladies.org/meetupr/}
  \item Report bugs at \url{https://github.com/rladies/meetupr/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Erin LeDell \email{oss@ledell.org}
  \item Olga Mierzwa-Sulima \email{olga@rladies.org}
  \item Lucy D'Agostino McGowan \email{lucy@rladies.org}
  \item Claudia Vitolo
}

Other contributors:
\itemize{
  \item Gabriela De Queiroz \email{gabriela@rladies.org} [contributor]
  \item Michael Beigelmacher [contributor]
  \item Augustina Ragwitz \email{augustina.ragwitz@ibm.com} [contributor]
  \item Greg Sutcliffe \email{github@emeraldverie.org} [contributor]
  \item Rick Pack \email{rickeyhp@gmail.com} [contributor]
  \item Ben Ubah \email{ubah.ben22@gmail.com} [contributor]
  \item Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID}) [contributor]
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID}) [contributor]
  \item R-Ladies Global [copyright holder]
}

}
\keyword{internal}
