% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetupr_auth}
\alias{meetupr_auth}
\title{Authenticate and return the current user}
\usage{
meetupr_auth(client_name = get_client_name())
}
\arguments{
\item{client_name}{OAuth client name}
}
\value{
A user list (invisibly) or \code{NULL} on failure
}
\description{
Attempts to retrieve information about the currently authenticated
user. On success a message is emitted and the user object is
returned. On failure a message is shown and \code{NULL} is returned.
}
