% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetupr_client}
\alias{meetupr_client}
\title{Create a Meetup OAuth client}
\usage{
meetupr_client(
  client_key = NULL,
  client_secret = NULL,
  client_name = get_client_name(),
  ...
)
}
\arguments{
\item{client_key}{Optional. The OAuth client ID.}

\item{client_secret}{Optional. The OAuth client secret.}

\item{client_name}{A string representing the name of the client. By
default, it is set to \code{"meetupr"} and retrieved from the
\code{MEETUPR_CLIENT_NAME} environment variable.}

\item{...}{Additional arguments passed to \code{httr2::oauth_client()}.}
}
\value{
An httr2 OAuth client object.
}
\description{
This function creates an httr2 OAuth client for the Meetup API.
It uses environment variables or built-in credentials as fallback.
}
