% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetupr_deauth}
\alias{meetupr_deauth}
\title{Deauthorize and remove cached authentication}
\usage{
meetupr_deauth(client_name = get_client_name())
}
\arguments{
\item{client_name}{OAuth client name}
}
\description{
Remove cached OAuth tokens and optionally clear stored credentials.
}
