% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{meetupr_sitrep}
\alias{meetupr_sitrep}
\title{Show meetupr authentication status}
\usage{
meetupr_sitrep()
}
\value{
Invisibly returns a list with authentication status
details.
}
\description{
This function checks the authentication status for the Meetup API.
It provides feedback on whether credentials are configured
correctly, tests API connectivity, and shows available
authentication methods.
}
\examples{
meetupr_sitrep()

}
