% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meetupr_verbose}
\alias{meetupr_verbose}
\title{Default verbose value}
\usage{
meetupr_verbose()
}
\value{
Result of \code{"meetupr.verbose"} option. If no \code{"meetupr.verbose"} option is set, \code{rlang::is_interactive()} is returned.
}
\description{
Wrapper around\if{html}{\out{<div class="sourceCode r">}}\preformatted{getOption("meetupr.verbose", rlang::is_interactive())
}\if{html}{\out{</div>}}
}
\keyword{internal}
