% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_event_attendance.R
\name{post_event_attendance}
\alias{post_event_attendance}
\title{Post the attendance for a specified event}
\usage{
post_event_attendance(urlname, event_id, member_id, status, api_key = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{event_id}{Character. The id of the event. Event ids can be obtained
using \code{\link[=get_events]{get_events()}} or by looking at the event page URL.}

\item{member_id}{Integer. The id of the member(s). Can be obtained using
\code{\link[=get_event_rsvps]{get_event_rsvps()}}. Can accept multiple members using \code{c()}.}

\item{status}{Character. Must be one of 'noshow', 'absent', 'attended'.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\description{
Post the attendance for a specified event
}
\references{
\url{https://www.meetup.com/meetup_api/docs/:urlname/events/:id/attendance/#edit}
}
