% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_group_response.R
\name{process_group_response}
\alias{process_group_response}
\title{Converts the response objects obtained by the get_data_for_group function
into a list of data.frames. Also merges the membership and profiles.}
\usage{
process_group_response(grl)
}
\arguments{
\item{grl}{A group response list generated by the get_data_for_group function}
}
\value{
A list of data.frames of group, membership, rsvps, event info
}
\description{
Converts the response objects obtained by the get_data_for_group function
into a list of data.frames. Also merges the membership and profiles.
}

