% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_auth.R
\name{set_api_key}
\alias{set_api_key}
\title{Store a legacy API key in the .state environment}
\usage{
set_api_key(x = NULL)
}
\description{
Store a legacy API key in the .state environment
}
\keyword{internal}
