% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_vector.R
\name{split_vector}
\alias{split_vector}
\title{Splits a vector into a groups of a specific number}
\usage{
split_vector(v, s)
}
\arguments{
\item{v}{A vector to split}

\item{s}{The number of elements per group}
}
\description{
Splits a vector into a groups of a specific number
}

