% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-groups.R
\name{find_groups}
\alias{find_groups}
\title{Find groups using text-based search}
\usage{
find_groups(query, ..., extra_graphql = NULL)
}
\arguments{
\item{query}{Character string to search for groups}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with group information
}
\description{
Find groups using text-based search
}
\examples{
\dontrun{
groups <- find_groups("R-Ladies")
}
}
