% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_groups.R
\name{find_groups}
\alias{find_groups}
\title{Find meetup groups matching a search query}
\usage{
find_groups(
  query,
  ...,
  topic_category_id = NULL,
  lat = 0,
  lon = 0,
  radius = 1e+08,
  extra_graphql = NULL,
  token = meetup_token()
)
}
\arguments{
\item{query}{Required search text}

\item{...}{Should be empty. Used for parameter expansion}
}
\description{
Find meetup groups matching a search query
}
