% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{find_groups}
\alias{find_groups}
\title{Find meetup groups using text-based search}
\usage{
find_groups(text = NULL, radius = "global", api_key = NULL)
}
\arguments{
\item{text}{string used to filter groups}

\item{radius}{can be either "global" (default) or distance in miles in the
range 0-100.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item name
\item urlname
\item who
\item members
\item status
\item created
\item organizer
\item lat
\item lon
\item city
\item country
\item timezone
}
}
\description{
Find meetup groups using text-based search
}
\examples{
\dontrun{
api_key <- Sys.getenv("rladies_api_key")
groups <- find_groups(text = "r-ladies", api_key = api_key)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/find/groups/}
}
