% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-groups.R
\name{find_groups}
\alias{find_groups}
\title{Find groups using text-based search}
\usage{
find_groups(query, max_results = 200, ..., extra_graphql = NULL)
}
\arguments{
\item{query}{Character string to search for groups}

\item{max_results}{Maximum number of results to return. Default: 200}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with group information
}
\description{
Find groups using text-based search
}
\examples{
\dontshow{
vcr::insert_example_cassette("find_groups", package = "meetupr")
meetupr:::mock_if_no_auth()
}
groups <- find_groups("R-Ladies")
\dontshow{
vcr::eject_cassette()
}
}
