% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_groups.R
\name{find_groups}
\alias{find_groups}
\title{Find meetup groups matching a search query}
\usage{
find_groups(text = NULL, topic_id = NULL, radius = "global",
  api_key = NULL)
}
\arguments{
\item{text}{Character. Raw full text search query.}

\item{topic_id}{Integer. Meetup.com topic ID.}

\item{radius}{can be either "global" (default) or distance in miles in the
range 0-100.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item urlname
\item created
\item members
\item status
\item organizer
\item lat
\item lon
\item city
\item state
\item country
\item timezone
\item organizer_id
\item organizer_name
\item category_id
\item category_name
\item resource
}
}
\description{
Find meetup groups matching a search query
}
\examples{
\dontrun{
api_key <- Sys.getenv("MEETUP_KEY")
groups <- find_groups(text = "r-ladies", api_key = api_key)
groups <- find_groups(topic_id = 1513883, api_key = api_key)  
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/find/groups/}
}
