% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topics.R
\name{find_topics}
\alias{find_topics}
\title{Find meetup topic IDs matching a text search query}
\usage{
find_topics(text = NULL, api_key = NULL)
}
\arguments{
\item{text}{Character. Raw full text search query.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item urlkey
\item description
\item member_count
\item group_count
\item resource
}
}
\description{
Find meetup topic IDs matching a text search query
}
\examples{
\dontrun{
api_key <- Sys.getenv("MEETUP_KEY")
topics <- find_topics(text = "R-Ladies", api_key = api_key)
# Note that R-Ladies has topic id 1513883
groups <- find_groups(topic_id = 1513883, api_key = api_key)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/find/groups/}
\url{https://www.meetup.com/meetup_api/docs/find/topics/}
}
