% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boards.R
\name{get_boards}
\alias{get_boards}
\title{Get the boards from a meetup group}
\usage{
get_boards(urlname, api_key)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}
}
\value{
List containing requested events.
}
\description{
Get the boards from a meetup group
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
api_key <- Sys.getenv("rladies_api_key")
meetup_boards <- get_boards(urlname = urlname,
                      api_key = api_key)
}
}

