% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-event.R
\name{get_event}
\alias{get_event}
\title{Get information for a specified event}
\usage{
get_event(id, extra_graphql = NULL, ...)
}
\arguments{
\item{id}{Required event ID}

\item{extra_graphql}{A graphql object. Extra objects to return}

\item{...}{Should be empty. Used for parameter expansion}

\item{asis}{Return the raw API response as-is without processing}
}
\value{
A meetup_event object with information about the specified event
}
\description{
Get information for a specified event
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_event", package = "meetupr")
meetupr:::mock_if_no_auth()
}
event <- get_event(id = "103349942")
\dontshow{
vcr::eject_cassette()
}
}
