% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_attendees.R
\name{get_event_attendees}
\alias{get_event_attendees}
\title{Get the attendees for a specified event}
\usage{
get_event_attendees(id, ..., extra_graphql = NULL, token = meetup_token())
}
\arguments{
\item{id}{Required event ID}

\item{...}{Should be empty. Used for parameter expansion}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item url
\item photo
\item organized_group_count
}
}
\description{
Get the attendees for a specified event
}
\examples{
\dontrun{
attendees <- get_event_attendees(id = "103349942!chp")
}
}
\references{
\url{https://www.meetup.com/api/schema/#Event}
\url{https://www.meetup.com/api/schema/#Ticket}
\url{https://www.meetup.com/api/schema/#User}
}
