% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-event-attendees.R
\name{get_event_attendees}
\alias{get_event_attendees}
\title{Get the attendees for a specified event}
\usage{
get_event_attendees(id, ..., extra_graphql = NULL)
}
\arguments{
\item{id}{Required event ID}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item url
\item photo
\item organized_group_count
}
}
\description{
Get the attendees for a specified event
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_event_attendees", package = "meetupr")
meetupr:::mock_if_no_auth()
}
attendees <- get_event_attendees(id = "103349942")
\dontshow{
vcr::eject_cassette()
}
}
\references{
\url{https://www.meetup.com/api/schema/#Event}
\url{https://www.meetup.com/api/schema/#RSVP}
\url{https://www.meetup.com/api/schema/#Member}
}
