% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_attendees.R
\name{get_event_attendees}
\alias{get_event_attendees}
\title{Get the attendees for a specified event}
\usage{
get_event_attendees(urlname, event_id, api_key = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{event_id}{Character. The id of the event. Event ids can be obtained
using \code{\link[=get_events]{get_events()}} or by looking at the event page URL.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item status
\item resource
}
}
\description{
Get the attendees for a specified event
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
past_events <- get_events(urlname = urlname,
                      event_status = "past")
event_id <- past_events$id[1]  #first event for this group
attendees <- get_event_attendees(urlname, event_id)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/:urlname/events/:id/attendance/#list}
}
