% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-comments.R
\name{get_event_comments}
\alias{get_event_comments}
\title{Get the comments for a specified event}
\usage{
get_event_comments(id, ..., extra_graphql = NULL)
}
\arguments{
\item{id}{Required event ID}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with comments data
}
\description{
Get the comments for a specified event
}
\examples{
\dontrun{
comments <- get_event_comments(id = "103349942!chp")
}
}
\references{
\url{https://www.meetup.com/api/schema/#Event}
\url{https://www.meetup.com/api/schema/#Comment}
}
