% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_rsvps.R
\name{get_event_rsvps}
\alias{get_event_rsvps}
\title{Get the RSVPs for a specified event}
\usage{
get_event_rsvps(id, ..., extra_graphql = NULL, token = meetup_token())
}
\arguments{
\item{id}{Required event ID}

\item{...}{Should be empty. Used for parameter expansion}
}
\value{
A tibble with the following columns:
\itemize{
\item member_id
\item member_name
\item member_url
\item member_is_host
\item guests
\item response
\item event_id
\item event_title
\item event_url
\item created
\item updated
}
}
\description{
Get the RSVPs for a specified event
}
\examples{
\dontrun{
rsvps <- get_event_rsvps(id = "103349942!chp")
}
}
\references{
\url{https://www.meetup.com/api/schema/#Event}
\url{https://www.meetup.com/api/schema/#Ticket}
\url{https://www.meetup.com/api/schema/#User}
}
