% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-event-rsvps.R
\name{get_event_rsvps}
\alias{get_event_rsvps}
\title{Get the RSVPs for a specified event}
\usage{
get_event_rsvps(id, ..., extra_graphql = NULL)
}
\arguments{
\item{id}{Required event ID}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with the following columns:
\itemize{
\item rsvp_id
\item member_id
\item member_name
\item member_url
\item member_photo
\item guests_count
\item response
}
}
\description{
Get the RSVPs for a specified event
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_event_rsvps", package = "meetupr")
meetupr:::mock_if_no_auth()
}
rsvps <- get_event_rsvps(id = "103349942")
\dontshow{
vcr::eject_cassette()
}
}
\references{
\url{https://www.meetup.com/api/schema/#Event}
\url{https://www.meetup.com/api/schema/#RSVP}
\url{https://www.meetup.com/api/schema/#Member}
}
