% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_rsvps.R
\name{get_event_rsvps}
\alias{get_event_rsvps}
\title{Get the RSVPs for a specified event}
\usage{
get_event_rsvps(urlname, event_id, verbose = meetupr_verbose())
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{event_id}{Character. The id of the event. Event ids can be obtained
using \code{\link[=get_events]{get_events()}} or by looking at the event page URL.}

\item{verbose}{logical; do you want informative messages? \code{TRUE} by default in
interactive sessions. Can be toggled by the \code{meetupr.verbose} option.}
}
\value{
A tibble with the following columns:
\itemize{
\item member_id
\item member_name
\item member_is_host
\item response
\item guests
\item created
\item updated
\item resource
}
}
\description{
Get the RSVPs for a specified event
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
upcoming_events <- get_events(urlname = urlname,
                      event_status = "past")
event_id <- upcoming_events$id[1]  #first event for this group
rsvps <- get_event_rsvps(urlname, event_id)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/:urlname/events/:event_id/rsvps/#list}
}
