% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get the events from a meetup group}
\usage{
get_events(urlname, ..., extra_graphql = NULL, token = meetup_token())
}
\arguments{
\item{urlname}{Required urlname of the Meetup group}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}

\item{token}{Meetup token}
}
\description{
Get the events from a meetup group
}
