% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get the events from a meetup group}
\usage{
get_events(urlname, api_key, event_status = NULL)
}
\arguments{
\item{api_key}{Your api key.}

\item{group_name}{The name of the group.}
}
\value{
The events.
}
\description{
Get the events from a meetup group
}
\examples{
get_events
}

