% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get the events from a meetup group}
\usage{
get_events(urlname, event_status = "upcoming", api_key = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicicated in the
\url{https://meetup.com} url.}

\item{event_status}{Character. Event type - defaults to "upcomming".
Valid inputs are:
\itemize{
\item cancelled
\item draft
\item past
\item proposed
\item suggested
\item upcoming
}}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item n_rsvp
\item time
\item events_resource
}
}
\description{
Get the events from a meetup group
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
past_events <- get_events(urlname = urlname,
                      event_status = "past")
upcoming_events <- get_events(urlname = urlname,
                      event_status = "upcoming")
}
}
