% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get the events from a meetup group}
\usage{
get_events(urlname, api_key, event_status = NULL)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}
\item{event_status}{A character string defining an event type. If empty upcoming meetups will be provided.}

\item{group_name}{The name of the group.}
\item{event_status}{Status of the event.}
}
\value{
List containing requested events.
}
\description{
Get the events from a meetup group
}

