% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-group.R
\name{get_group}
\alias{get_group}
\title{Get detailed information about a Meetup group}
\usage{
get_group(urlname)
}
\arguments{
\item{urlname}{The URL name of the Meetup group (e.g., "rladies-lagos")}
}
\value{
A list containing detailed information about the Meetup group
}
\description{
Get detailed information about a Meetup group
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_group", package = "meetupr")
meetupr:::mock_if_no_auth()
}
get_group("rladies-lagos")
\dontshow{
vcr::eject_cassette()
}
}
