% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get list of meetup groups using text-based search}
\usage{
get_groups(api_key, ...)
}
\arguments{
\item{api_key}{Your api key.}

\item{...}{any parameter listed in meetup API (\url{https://www.meetup.com/meetup_api/docs/find/groups/})}
}
\value{
List containing relevant groups.
}
\description{
Get list of meetup groups using text-based search
}
\examples{
\dontrun{
api_key <- Sys.getenv("rladies_api_key")
groups <- get_groups(api_key = api_key,
                     sign   = "true",
                     text = "r-ladies",
                     radius = "global",
                     only = "urlname")
rladies_groups <- as.character(unlist(groups))
}
}
