% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get list of meetup groups using text-based search}
\usage{
get_groups(api_key, text = NULL, radius = "global")
}
\arguments{
\item{api_key}{Your api key.}

\item{text}{string used to filter groups}

\item{radius}{can be either "global" (default) or distance in miles in the
range 0-100.}
}
\value{
Tibble containing the list of selected groups. The table contains 12
columns: name, urlname, who, members, status, created, organizer, lat, lon,
city, country, timezone.
}
\description{
Get list of meetup groups using text-based search
}
\examples{
\dontrun{
api_key <- Sys.getenv("rladies_api_key")
groups <- get_groups(api_key = api_key, text = "r-ladies")
}
}
