% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_meetup_attendees}
\alias{get_meetup_attendees}
\title{DEPRECATED: Get the attendees}
\usage{
get_meetup_attendees(urlname, api_key, event_id)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}

\item{event_id}{The id of the event.}
}
\description{
DEPRECATED: Get the attendees
}
\examples{
\dontrun{
urlname <- "rladies-san-francisco"
api_key <- Sys.getenv("MEETUP_KEY")
past_events <- get_events(urlname = urlname,
                      api_key = api_key,
                      event_status = "past")
event_id <- past_events[[1]]$id
attendees <- get_meetup_attendees(urlname, api_key, event_id)
}
}
