% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetup_attendees.R
\name{get_meetup_attendees}
\alias{get_meetup_attendees}
\title{Get the attendees}
\usage{
get_meetup_attendees(group_name, id, api_key)
}
\arguments{
\item{group_name}{The name of the group.}

\item{id}{The id of the event.}

\item{api_key}{Your api key.}
}
\description{
Get the attendees
}

