% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetup_comments.R
\name{get_meetup_comments}
\alias{get_meetup_comments}
\title{Get the comments}
\usage{
get_meetup_comments(urlname, api_key, event_id)
}
\arguments{
\item{urlname}{The name of the group.}

\item{api_key}{Your api key.}

\item{event_id}{The id of the event.}
}
\description{
Get the comments
}

