% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-members.R
\name{get_members}
\alias{get_members}
\title{Get the members from a meetup group}
\usage{
get_members(urlname, ..., extra_graphql = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item member_url
\item photo_link
\item status
\item role
\item created
\item most_recent_visit
}
}
\description{
Get the members from a meetup group
}
