% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Get the current meetup members from a meetup group}
\usage{
get_members(urlname, ..., extra_graphql = NULL, token = meetup_token())
}
\arguments{
\item{urlname}{Required urlname of the Meetup group}

\item{...}{Should be empty. Used for parameter expansion}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item member_url
\item photo_link
\item status
\item role
\item created
\item most_recent_visit
}
}
\description{
Get the current meetup members from a meetup group
}
\examples{
\dontrun{
members <- get_members("rladies-remote")
}
}
\references{
\url{https://www.meetup.com/api/schema/#GroupMembership}
\url{https://www.meetup.com/api/schema/#User}
}
