% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-members.R
\name{get_members}
\alias{get_members}
\title{Get the members from a meetup group}
\usage{
get_members(urlname, max_results = 200, ..., extra_graphql = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{max_results}{Maximum number of results to return. Default: 200}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}
}
\value{
A tibble with group members
}
\description{
Get the members from a meetup group
}
\examples{
\dontshow{
vcr::insert_example_cassette("get_members", package = "meetupr")
meetupr:::mock_if_no_auth()
}
get_members("rladies-lagos")
\dontshow{
vcr::eject_cassette()
}
}
