% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Get the current meetup members}
\usage{
get_members(urlname, api_key = NULL)
}
\arguments{
\item{urlname}{Character. The name of the group as indicicated in the
\url{https://meetup.com} url.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item status
\item photo_link
\item resource
}
}
\description{
Get the current meetup members
}
\examples{
\dontrun{
urlname <- "rladies-nashville"
members <- get_members(urlname)
}
}
