% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{get_pro_groups}
\alias{get_pro_groups}
\title{Get the groups under a meetup pro organisation}
\usage{
get_pro_groups(urlname)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item status
\item founded
\item member_count
\item member_count
\item upcoming_events
\item city
\item country
\item state
\item lat
\item lon
\item urlname
}
}
\description{
Get the groups under a meetup pro organisation
}
\examples{
\dontrun{
urlname <- "rladies"
members <- get_pro_groups(urlname)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/pro/:urlname/groups/}
}
