% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{get_pro_groups}
\alias{get_pro_groups}
\title{Get the current meetup members from a meetup group}
\usage{
get_pro_groups(urlname, api_key = NULL, verbose = TRUE)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{api_key}{Character. Your api key. Defaults to checking your environment
for a parameter called "MEETUP_KEY" via \code{Sys.getenv("MEETUP_KEY")}. This key can be
obtained from \url{https://secure.meetup.com/meetup_api/key/}. To set the environment variable
run \code{Sys.setenv(MEETUP_KEY = "PASTE YOUR MEETUP KEY HERE")}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
A tibble with the following columns:
\itemize{
\item id
\item name
\item status
\item founded
\item members
\item upcoming_events
\item past events
\item city
\item country
\item state
\item lat
\item lon
\item urlname
}
}
\description{
Get the current meetup members from a meetup group
}
\examples{
\dontrun{
urlname <- "rladies"
members <- get_pro_groups(urlname)
}
}
\references{
\url{https://www.meetup.com/meetup_api/docs/pro/:urlname/groups/}
}
