% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_auth.R
\name{meetup_auth}
\alias{meetup_auth}
\alias{meetup_token}
\alias{meetup_deauth}
\alias{meetup_token_path}
\title{Authorize \code{meetupr}}
\usage{
meetup_auth(
  token = meetup_token_path(),
  new_user = FALSE,
  key = getOption("meetupr.consumer_key"),
  secret = getOption("meetupr.consumer_secret"),
  cache = getOption("meetupr.httr_oauth_cache"),
  verbose = TRUE,
  set_renv = TRUE,
  token_path = NULL
)

meetup_token(verbose = FALSE)

meetup_deauth(clear_cache = TRUE, verbose = TRUE)

meetup_token_path()
}
\arguments{
\item{token}{optional; an actual token object or the path to a valid token
stored as an \code{.rds} file}

\item{new_user}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
want to wipe the slate clean and re-authenticate with the same or different
Meetup account. This disables the \code{.httr-oauth} file in current
working directory.}

\item{key, secret}{the "Client ID" and "Client secret" for the application;
defaults to the ID and secret built into the \code{meetupr} package}

\item{cache}{logical indicating if \code{meetupr} should cache
credentials in the default cache file \code{.httr-oauth}}

\item{verbose}{logical; do you want informative messages?}

\item{set_renv}{Logical indicating whether to save the created token
as the default environment meetup token variable. Defaults to TRUE,
meaning the token is saved to user's home directory as either the user
provided path, or
".meetup_token.rds" (or, if that already exists, then
.meetup_token1.rds or .meetup_token2.rds, etc.) and the path to the
token to said token is then set in the user's .Renviron file and re-
read to start being used in current active session.
If \code{cache} is \code{FALSE} this is ignored.}

\item{token_path}{Path where to save the token. If \code{set_renv} is \code{FALSE},
this is ignored.}

\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}
}
\value{
a \code{request} object (an S3 class provided by \code{httr})

Either NULL or the path in which the token is saved.
}
\description{
Authorize \code{meetupr} via the OAuth API. You will be directed to a web
browser, asked to sign in to your Meetup account, and to grant \code{meetupr}
permission to operate on your behalf. By default, these user credentials are
saved to a file in your home directory whose path is saved in \code{.Renviron}
as \code{MEETUPR_PAT}.
If you set \code{set_env} to \code{FALSE} but \code{cache} to \code{TRUE},
they are cached in a file named \code{.httr-oauth} in the current working directory.

If token is not already available, call \code{\link{meetup_auth}} to either
load from cache or initiate OAuth2.0 flow. Return the token -- not "bare"
but, rather, prepared for inclusion in downstream requests.

Suspend \code{\link{meetupr}}'s authorization to place requests to the Meetup
APIs on behalf of the authenticated user.
}
\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
meetup_auth()

## store token in an object and then to file
ttt <- meetup_auth()
saveRDS(ttt, "ttt.rds")

## load a pre-existing token
meetup_auth(token = ttt)       # from an object
meetup_auth(token = "ttt.rds") # from .rds file
}
\dontrun{
meetup_token()
}
\dontrun{
meetup_deauth()
}
meetup_token_path()
}
\concept{auth functions}
\keyword{internal}
