% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_auth.R
\name{meetup_deauth}
\alias{meetup_deauth}
\title{Suspend authorization}
\usage{
meetup_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Suspend \code{\link{meetupr}}'s authorization to place requests to the Meetup
APIs on behalf of the authenticated user.
}
\examples{
\dontrun{
meetup_deauth()
}
}
\seealso{
Other auth functions: 
\code{\link{meetup_auth}()},
\code{\link{meetup_token}()}
}
\concept{auth functions}
