% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{meetup_pro}
\alias{meetup_pro}
\alias{get_pro_groups}
\alias{get_pro_events}
\title{Meetup pro functions}
\usage{
get_pro_groups(urlname, ..., extra_graphql = NULL, token = meetup_token())

get_pro_events(
  urlname,
  status = NULL,
  ...,
  extra_graphql = NULL,
  token = meetup_token()
)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}

\item{token}{Meetup token}

\item{status}{Which status the events should have.}
}
\value{
A tibble with meetup pro information
}
\description{
The pro functions only work if the querying users
had a meetup pro account.
}
\section{Functions}{
\itemize{
\item \code{get_pro_groups}: retrieve groups in a pro network

\item \code{get_pro_events}: retrieve events from a pro network
}}

\examples{
\dontrun{
urlname <- "rladies"
members <- get_pro_groups(urlname)

past_events <- get_pro_events(urlname = urlname,
                      status = "PAST")
upcoming_events <- get_pro_events(urlname = urlname,
                      status = "UPCOMING")
all_events <- get_pro_events(urlname = urlname)
}
}
\references{
\url{https://www.meetup.com/api/schema/#ProNetwork}
}
