% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{get_pro_groups}
\alias{get_pro_groups}
\alias{get_pro_events}
\title{#' Meetup pro functions
#'
#' The pro functions only work if the querying users
#' had a meetup pro account.
#'
#' \describe{
#'   \item{get_pro_groups}{Get the current meetup members from a pro meetup group}
#'   \item{get_pro_events}{Get pro group events}
#' }
#'
#' @template urlname
#'
#' @references
#' \url{https://www.meetup.com/api/schema/#ProNetwork}
#'
#' @examples
#' \dontrun{
#' urlname <- "rladies"
#' members <- get_pro_groups(urlname)
#'
#' past_events <- get_events(urlname = urlname,
#'                       status = "PAST")
#' upcoming_events <- get_events(urlname = urlname,
#'                       status = "UPCOMING")
#' all_events <- get_events(urlname = urlname)
#'}
#'
#' @return A tibble with meetup information}
\usage{
get_pro_groups(urlname, ..., extra_graphql = NULL, token = meetup_token())

get_pro_events(
  urlname,
  status = NULL,
  ...,
  extra_graphql = NULL,
  token = meetup_token()
)
}
\value{
A tibble with the meetup pro group information
}
\description{
#' Meetup pro functions
#'
#' The pro functions only work if the querying users
#' had a meetup pro account.
#'
#' \describe{
#'   \item{get_pro_groups}{Get the current meetup members from a pro meetup group}
#'   \item{get_pro_events}{Get pro group events}
#' }
#'
#' @template urlname
#'
#' @references
#' \url{https://www.meetup.com/api/schema/#ProNetwork}
#'
#' @examples
#' \dontrun{
#' urlname <- "rladies"
#' members <- get_pro_groups(urlname)
#'
#' past_events <- get_events(urlname = urlname,
#'                       status = "PAST")
#' upcoming_events <- get_events(urlname = urlname,
#'                       status = "UPCOMING")
#' all_events <- get_events(urlname = urlname)
#'}
#'
#' @return A tibble with meetup information
}
