% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pro.R
\name{meetup_pro}
\alias{meetup_pro}
\alias{get_pro_groups}
\alias{get_pro_events}
\title{Retrieve information about Meetup Pro networks, including groups and events.}
\usage{
get_pro_groups(urlname, max_results = NULL, ..., extra_graphql = NULL)

get_pro_events(
  urlname,
  status = NULL,
  max_results = NULL,
  ...,
  extra_graphql = NULL
)
}
\arguments{
\item{urlname}{Character. The name of the group as indicated in the
\url{https://meetup.com} url.}

\item{max_results}{Maximum number of results to return. Default: 200}

\item{...}{Should be empty. Used for parameter expansion}

\item{extra_graphql}{A graphql object. Extra objects to return}

\item{status}{Which status the events should have.}
}
\value{
A tibble with meetup pro information
}
\description{
Meetup Pro is a premium service for organizations managing multiple Meetup groups.
This functionality allows you to access details about the groups within a Pro network
and the events they host.
}
\section{Functions}{
\itemize{
\item \code{get_pro_groups()}: retrieve groups in a pro network

\item \code{get_pro_events()}: retrieve events from a pro network

}}
\examples{
\dontshow{
vcr::insert_example_cassette("get_pro", package = "meetupr")
meetupr:::mock_if_no_auth()
}
urlname <- "rladies"
members <- get_pro_groups(urlname)

upcoming_events <- get_pro_events(urlname, "upcoming", max_results = 5)
\dontshow{
vcr::eject_cassette()
}
}
\references{
\url{https://www.meetup.com/api/schema/#ProNetwork}
}
