% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetup_query}
\alias{meetup_query}
\title{Execute GraphQL query}
\usage{
meetup_query(.query, ...)
}
\arguments{
\item{.query}{GraphQL query string}

\item{...}{Variables to pass to query}
}
\description{
Execute GraphQL query
}
