% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_query.R
\name{meetup_query}
\alias{meetup_query}
\title{General Meetup query}
\usage{
meetup_query(api_path, ...)
}
\arguments{
\item{api_path}{url path to send the meetup query after the 'https://api.meetup.com/'}

\item{...}{additional parameters to the query}
}
\value{
raw list results from the query, not tidied
}
\description{
Create your own query to the Meetup API
(https://www.meetup.com/meetup_api/docs/).
This function is convenient if you want to create
a query to other Meetup API endpoints that we
have not created functions for. If you do create
a function that quieries a currently unsupported
endpoint and a results tidier for this, we are happy
to receive PR's with your suggestion.
}
\examples{
\dontrun{

meetup_query("rladies-oslo")

meetup_query("rladies-oslo/events", event_status = "past")

}
}
