% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetupr-package.R
\docType{package}
\name{meetupr-package}
\alias{meetupr}
\alias{meetupr-package}
\title{meetupr: Meetup R API}
\description{
Provides access to data from 'meetup.com' (see
    https://www.meetup.com/meetup_api/ for more information).
}
\author{
\strong{Maintainer}: Erin LeDell \email{oss@ledell.org}

Authors:
\itemize{
  \item Olga Mierzwa-Sulima \email{olga@rladies.org}
  \item Lucy D'Agostino McGowan \email{lucy@rladies.org}
  \item Claudia Vitolo
}

Other contributors:
\itemize{
  \item Gabriela De Queiroz \email{gabriela@rladies.org} [contributor]
  \item Michael Beigelmacher [contributor]
  \item Augustina Ragwitz \email{augustina.ragwitz@ibm.com} [contributor]
  \item Greg Sutcliffe \email{github@emeraldverie.org} [contributor]
  \item Rick Pack \email{rickeyhp@gmail.com} [contributor]
  \item Ben Ubah \email{ubah.ben22@gmail.com} [contributor]
}

}
\keyword{internal}
