% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{meetupr_verbose}
\alias{meetupr_verbose}
\title{Get/set verbose option}
\usage{
meetupr_verbose(verbose = NULL)
}
\arguments{
\item{verbose}{Logical for verbose output}
}
\description{
Get/set verbose option
}
