% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pro.R
\name{slowly_get_events}
\alias{slowly_get_events}
\title{to avoid making too many
requests too rapidly when
getting pro events}
\usage{
slowly_get_events(...)
}
\arguments{
\item{...}{arguments to get_events}
}
\description{
to avoid making too many
requests too rapidly when
getting pro events
}
